arborescence & formatting
