+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/adapters/bff/outputs/HealthcheckDataOutput.kt
@@ -2,7 +2,7 @@ package fr.gouv.cacem.monitorenv.infrastructure.api.adapters.bff.outputs
 
 import fr.gouv.cacem.monitorenv.domain.entities.health.Health
 
-data class HealthDataOutput(
+data class HealthcheckDataOutput(
     val numberOfRegulatoryAreas: Long,
     val numberOfMissions: Long,
     val numberOfNatinfs: Long,
@@ -10,7 +10,7 @@ data class HealthDataOutput(
     val numberOfReportings: Long,
 ) {
     companion object {
-        fun fromHealth(health: Health) = HealthDataOutput(
+        fun fromHealth(health: Health) = HealthcheckDataOutput(
             numberOfRegulatoryAreas = health.numberOfRegulatoryAreas,
             numberOfMissions = health.numberOfMissions,
             numberOfNatinfs = health.numberOfNatinfs,
