+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/endpoints/Spa.kt
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping
  * This controller redirects all unhandled routes by the backend to our SPA frontend
  */
 @Controller
-class SpaController : ErrorController {
+class Spa : ErrorController {
     @RequestMapping("/error")
     fun error(request: HttpServletRequest, response: HttpServletResponse): Any {
         response.status = HttpStatus.OK.value()
