+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/endpoints/publicapi/v1/Healthcheck.kt
@@ -1,7 +1,7 @@
 package fr.gouv.cacem.monitorenv.infrastructure.api.endpoints.publicapi.v1
 
 import fr.gouv.cacem.monitorenv.domain.use_cases.healthcheck.GetHealthcheck
-import fr.gouv.cacem.monitorenv.infrastructure.api.adapters.bff.outputs.HealthDataOutput
+import fr.gouv.cacem.monitorenv.infrastructure.api.adapters.bff.outputs.HealthcheckDataOutput
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.web.bind.annotation.GetMapping
@@ -17,7 +17,7 @@ class Healthcheck(
 
     @GetMapping("")
     @Operation(summary = "Get healtcheck for all resources")
-    fun getHealthcheck(): HealthDataOutput {
-        return HealthDataOutput.fromHealth(getHealthcheck.execute())
+    fun getHealthcheck(): HealthcheckDataOutput {
+        return HealthcheckDataOutput.fromHealth(getHealthcheck.execute())
     }
 }
